[FIX] account: reconciliation popover with lines detail was empty
commit 8c5fea6d525864ca0e735841b751e2c253174212 update our bootstrap version to 4.3.1, one change of that version is that
bootstrap now escape the content of popover and title removing the table element. https://blog.getbootstrap.com/2019/02/13/bootstrap-4-3-1-and-3-4-1/
This result in the popover in the reconciliation widget being empty.

Therefore since the content of the popover was safe as it was rendered by Qweb.render, we disabled completely the sanitizer
for the popover in the reconciliation widget.

closes odoo/odoo#31797

Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>