[FIX] packaging: include pot files
Since #26134 the po files are compared with pot content when loading a
translation. If a term is found multiple times in a po file, the terms
that does not exists in the pot files are ignored. If the pot files does
not exists, the unicity constraint leads to an error.

Closes #31583

closes odoo/odoo#31823

Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>