[FIX] website_event: wait drag and drop to be finished
During the event tour, the save button is clicked before the drag and
drop is finished and cause an uncaught exception.

With this commit, an extra trigger is used to check the presence of the
dirty flag that is set when the page is changed and is ready to be
saved.